eliminates the need for a 'removes' param
The FaithComesByHearingRenamer used to use a `renames` parameter so that
it could know which part of each filename to remove.  However, this is
as easily accomplished by a relatively simple regex homed in the renamer
itself, so the need to have that parameter for each resource from Faith
Comes By Hearing is not necessary.